add feature for thermo rawfile support
Note that this feauture needs to be called thermorawfilereader. mzdata's
feature is called thermo, but the checks, including the one in the
`mz_read` macro tests for thermorawfilereader. The feature here needs to
match the one tested for in the macro, otherwise rawfiles are detected
as a file type, but the macro still wouldn't support it.