update mzdata
This leaves us with 1 test failure: in our test MGF, PEPMASS is preceded
by CHARGE. mzdata now parses the charge, but implicitly assumes there is
a PEPMASS first. When that isn't the case, there is no precursor ion to
assign the charge value to, resulting in a failure of our unit test.